refactor linear neck
